No more crashes :'(

+++ GameLogic/Functions.h
@@ -7,7 +7,7 @@ class Functions {
 		static bool Intersect(Position start, Position end, float z=1.25f);
 		static float GetDepth(Position pos, float height);
 		static Position ProjectPos(Position pos, float height=2.0f);
-		static unsigned long GetPlayerGuid();
+		static unsigned long long GetPlayerGuid();
 		static int GetPositionCircle();
 		static void EnumerateVisibleObjects(int filter);
 		static uintptr_t GetObjectPtr(unsigned long long guid);
